IJPL-214442: Use Byte instead of Int

+++ kotlinx-coroutines-core/common/src/internal/NamedSoftParallelismDispatcher.kt
@@ -37,7 +37,7 @@ internal class NamedSoftParallelismDispatcher<D>(
         return SoftLimitedDispatcher(this, parallelism, name)
     }
 
-    override fun tryAdjustParallelism(parallelismDelta: Int): Int {
+    override fun tryAdjustParallelism(parallelismDelta: Byte): Byte {
         return dispatcher.tryAdjustParallelism(parallelismDelta)
     }
 }
